move form field generation code out to a helper
And replace the code in the UpdatePerson, NewPerson, Areas and
ConstituencyDetail views with a call to the helper
